requirements.txt
aioredis==1.3.1
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
coverage==5.4
cryptography==3.4.7
daphne==3.0.2
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-guardian==2.3.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.13
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
pytz==2020.5
redis==3.5.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
